hubolib: Pick layout per output format
